+++ src/components/Posts/PostList/PostListItem/PostListItem.stories.tsx
@@ -1,25 +1,52 @@
-import { ComponentStory, ComponentMeta } from "@storybook/react";
+// import { ComponentStory, ComponentMeta } from "@storybook/react";
+// import type {Meta} from "@storybook/react";
+
+// import Component from "./PostListItem";
+
+// export default {
+//   title: "Blogs & Webinars/Post List/PostListItem",
+//   component: Component,
+// } as ComponentMeta<typeof Component>;
+
+// const Template: Meta<typeof Component> = (args) => {
+//   return <Component {...args} />;
+// };
+
+// export const PostListItem = Template.bind({});
+// PostListItem.args = {
+//   titleTag: "h3",
+//   title: "The long and winding road",
+//   summary:
+//     "The long and winding road  That leads to your door  Will never disappear  I've seen that road before  It always leads me here  Lead me to you door  The wild and windy night  That the rain washed away  Has left a pool of tears  Crying for the day  Why leave me standing here?  Let me know the way",
+//   slug: "the-long-and-winding-road",
+//   category: { title: "Oak updates", slug: "oak-updates" },
+//   date: "2022-10-04",
+//   withImage: true,
+//   contentType: "blog-post",
+// };
+
+import type { Meta, StoryObj } from "@storybook/react";
 
 import Component from "./PostListItem";
 
-export default {
+const meta: Meta<typeof Component> = {
   title: "Blogs & Webinars/Post List/PostListItem",
   component: Component,
-} as ComponentMeta<typeof Component>;
-
-const Template: ComponentStory<typeof Component> = (args) => {
-  return <Component {...args} />;
 };
 
-export const PostListItem = Template.bind({});
-PostListItem.args = {
-  titleTag: "h3",
-  title: "The long and winding road",
-  summary:
-    "The long and winding road  That leads to your door  Will never disappear  I've seen that road before  It always leads me here  Lead me to you door  The wild and windy night  That the rain washed away  Has left a pool of tears  Crying for the day  Why leave me standing here?  Let me know the way",
-  slug: "the-long-and-winding-road",
-  category: { title: "Oak updates", slug: "oak-updates" },
-  date: "2022-10-04",
-  withImage: true,
-  contentType: "blog-post",
+export default meta;
+type Story = StoryObj<typeof Component>;
+
+export const PostListItem: Story = {
+  args: {
+    titleTag: "h3",
+    title: "The long and winding road",
+    summary:
+      "The long and winding road  That leads to your door  Will never disappear  I've seen that road before  It always leads me here  Lead me to you door  The wild and windy night  That the rain washed away  Has left a pool of tears  Crying for the day  Why leave me standing here?  Let me know the way",
+    slug: "the-long-and-winding-road",
+    category: { title: "Oak updates", slug: "oak-updates" },
+    date: "2022-10-04",
+    withImage: true,
+    contentType: "blog-post",
+  },
 };