+++ src/components/UnitAndLessonLists/LessonList/LessonList.stories.tsx
@@ -2,6 +2,7 @@ import React from "react";
 import { ComponentStory, ComponentMeta } from "@storybook/react";
 
 import lessonListingFixture from "../../../node-lib/curriculum-api/fixtures/lessonListing.fixture";
+import AnalyticsDecorator from "../../../storybook-decorators/AnalyticsDecorator";
 
 import { LessonListProps } from "./LessonList";
 
@@ -9,6 +10,7 @@ import Component from ".";
 
 export default {
   title: "Lists/Lesson List",
+  decorators: [AnalyticsDecorator],
   component: Component,
 } as ComponentMeta<typeof Component>;
 
@@ -23,6 +25,8 @@ const currentPageItems: LessonListProps = {
   paginationProps: {
     currentPage: 1,
     totalPages: 2,
+    totalResults: 10,
+    pageSize: 5,
   },
   unitTitle: "Unit title",
 };