+++ src/node-lib/curriculum-api/index.ts
@@ -290,6 +290,8 @@ const programmesData = z.object({
   tierTitle: z.string().nullable().optional(),
 });
 
+const programmesArray = z.array(programmesData);
+
 const subjectListingData = z.object({
   keyStageSlug: z.string(),
   keyStageTitle: z.string(),
@@ -312,12 +314,13 @@ const unitListingData = z.object({
   subjectSlug: z.string(),
   subjectTitle: z.string(),
   tierSlug: z.string().nullable(),
+  totalUnitCount: z.number(),
   tiers: tiersData,
   units: unitsData,
   learningThemes: z.array(
     z.object({
-      learningThemeTitle: z.string(),
-      learningThemeSlug: z.string(),
+      learningThemeTitle: z.string().nullable(),
+      learningThemeSlug: z.string().nullable(),
     })
   ),
 });
@@ -368,7 +371,7 @@ const getFirstResultOrWarnOrFail =
     return firstResult;
   };
 
-const getFirstResultOrNull =
+export const getFirstResultOrNull =
   () =>
   <T>({ results }: { results: T[] }) => {
     const [firstResult] = results;
@@ -378,6 +381,18 @@ const getFirstResultOrNull =
 
     return firstResult;
   };
+
+export const filterOutDuplicateProgrammesOrNull = (
+  programmesAvailable: ProgrammesData[],
+  programmesUnavailable: ProgrammesData[]
+) => {
+  return programmesUnavailable.filter(
+    (unavailable) =>
+      !programmesAvailable.some(
+        (available) => available.subjectSlug === unavailable.subjectSlug
+      )
+  );
+};
 
 const curriculumApi = {
   searchPage: async () => {
@@ -400,11 +415,17 @@ const curriculumApi = {
 
     const keyStage = getFirstResultOrWarnOrFail()({ results: keyStages });
 
+    const filteredUnavailableProgrammeDuplicate =
+      filterOutDuplicateProgrammesOrNull(
+        programmesArray.parse(programmesAvailable),
+        programmesArray.parse(programmesUnavailable)
+      );
+
     return subjectListingData.parse({
       keyStageSlug: keyStage.slug,
       keyStageTitle: keyStage.title,
       programmesAvailable,
-      programmesUnavailable,
+      programmesUnavailable: filteredUnavailableProgrammeDuplicate || [],
     });
   },
   unitListingPaths: async () => {
@@ -419,34 +440,26 @@ const curriculumApi = {
     const { units = [], programmes = [], tiers = [] } = transformMVCase(res);
 
     const programme = getFirstResultOrWarnOrFail()({ results: programmes });
-    const learningThemes = units
-      ?.filter((unit) => unit?.themeSlug !== "no-theme")
-      .map((unitWithTheme) => ({
-        learningThemeSlug: unitWithTheme?.themeSlug || "",
-        learningThemeTitle: unitWithTheme?.themeTitle || "",
-      }))
-      .sort((a, b) => {
-        if (a.learningThemeTitle < b.learningThemeTitle) {
-          return -1;
-        }
-        if (a.learningThemeTitle > b.learningThemeTitle) {
-          return 1;
-        }
-        return 0;
-      });
+    const learningThemes = units.map((unitWithTheme) => ({
+      learningThemeSlug: unitWithTheme?.themeSlug,
+      learningThemeTitle: unitWithTheme?.themeTitle || "No theme",
+    }));
 
     // !Refactor index signature to be more specific
 
-    type LearningTheme = {
-      [key: string]: string;
-    };
-
-    const filteredDuplicatedLearningThemes = learningThemes.filter(
-      (learningTheme: LearningTheme, index, learningThemeToCompare) =>
-        learningThemeToCompare.findIndex((lt: LearningTheme) =>
-          ["learningThemeSlug"].every((l) => lt[l] === learningTheme[l])
-        ) === index
-    );
+    const filteredDuplicatedLearningThemes = [
+      ...new Map(
+        learningThemes.map((theme) => [JSON.stringify(theme), theme])
+      ).values(),
+    ].sort((a, b) => {
+      if (a.learningThemeTitle < b.learningThemeTitle) {
+        return -1;
+      }
+      if (a.learningThemeTitle > b.learningThemeTitle) {
+        return 1;
+      }
+      return 0;
+    });
 
     return unitListingData.parse({
       programmeSlug: programme?.programmeSlug,
@@ -454,6 +467,7 @@ const curriculumApi = {
       keyStageTitle: programme?.keyStageTitle,
       subjectSlug: programme?.subjectSlug,
       subjectTitle: programme?.subjectTitle,
+      totalUnitCount: programme?.totalUnitCount,
       tierSlug: programme?.tierSlug || null,
       learningThemes: filteredDuplicatedLearningThemes,
       tiers,