+++ src/pages/beta/teachers/programmes/[programmeSlug]/units/[unitSlug]/lessons/[lessonSlug].tsx
@@ -19,8 +19,6 @@ import { getSeoProps } from "../../../../../../../../browser-lib/seo/getSeoProps
 import Typography, {
   Heading,
   Hr,
-  LI,
-  UL,
 } from "../../../../../../../../components/Typography";
 import ButtonAsLink from "../../../../../../../../components/Button/ButtonAsLink";
 import Grid from "../../../../../../../../components/Grid";
@@ -43,6 +41,7 @@ import type {
   DownloadResourceButtonNameValueType,
 } from "../../../../../../../../browser-lib/avo/Avo";
 import useAnalyticsPageProps from "../../../../../../../../hooks/useAnalyticsPageProps";
+import LessonOverview from "../../../../../../../../components/LessonOverview/LessonOverview";
 
 export type LessonOverviewPageProps = {
   curriculumData: LessonOverviewData;
@@ -204,23 +203,6 @@ const LessonOverviewPage: NextPage<LessonOverviewPageProps> = ({
           </Box>
         ) : (
           <>
-            <Flex $flexDirection="column">
-              <Heading tag={"h2"} $font={"heading-6"} $mb={16}>
-                Core content
-              </Heading>
-              <UL $pl={28}>
-                {coreContent.map((contentString, i) => {
-                  if (!contentString) {
-                    return null;
-                  }
-                  return (
-                    <LI key={`core-content-string-${i}`} $font={"list-item-1"}>
-                      {contentString}
-                    </LI>
-                  );
-                })}
-              </UL>
-            </Flex>
             <Flex $mt={12} $flexWrap={"wrap"}>
               {hasDownloadableResources && (
                 <ButtonAsLink
@@ -262,7 +244,19 @@ const LessonOverviewPage: NextPage<LessonOverviewPageProps> = ({
             $mt={16}
           /> */}
             </Flex>
+
             <Hr $color={"oakGrey3"} />
+
+            {coreContent && (
+              <ExpandingContainer
+                title={"Lesson overview"}
+                downloadable={false}
+                toggleClosed={true}
+                {...curriculumData}
+              >
+                <LessonOverview coreContent={coreContent} />
+              </ExpandingContainer>
+            )}
             {presentationUrl && !hasCopyrightMaterial && (
               <ExpandingContainer
                 downloadable={true}